requirements.txt
flask
flask-sqlalchemy
flask-mysqldb
python-dotenv
wtforms
flask_wtf
flask_login
bcrypt
pytest